disks: simpler check whether disk test file exists

+++ plinth/modules/disks/tests/test_disks.py
@@ -53,11 +53,7 @@ class Disk():
         directory = os.path.dirname(os.path.realpath(__file__))
         disk_file = os.path.join(directory, 'temp_disk.img')
 
-        try:
-            os.stat(disk_file)
-        except FileNotFoundError:
-            pass
-        else:
+        if os.path.isfile(disk_file):
             os.remove(disk_file)
 
         command = 'dd if=/dev/zero of={file} bs=1M count={size}' \
